Move files in storage/Embedding* into storage/embedding
